chore(auto): move code to new package
