+++ RemootioAndroidAuto/automotive/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             android:name="androidx.car.app.minCarApiLevel"
             android:value="1" />
 
-        <service android:name="chamilton0.remootioandroidws.RemootioService"
+        <service android:name="chamilton0.remootioandroidws.auto.RemootioService"
             android:exported="true">
             <intent-filter>
                 <action android:name="androidx.car.app.CarAppService" />
